fix purecap bug

+++ src/slimguard.c
@@ -204,14 +204,14 @@ void *get_next(uint8_t index){
     /* We require slots managing power of two allocations to be aligned on
      * their sizes to properly serve memalign requests */
     if(!(Class[index].size % PAGE_SIZE)) {
-        uint64_t old_ret = (uint64_t)ret;
+        intptr_t old_ret = (intptr_t)ret;
 
-        ret = (void *)(intptr_t)(((uint64_t)ret + (uint64_t)(Class[index].size) - 1) &
-                ~((uint64_t)(Class[index].size) - 1));
-        Class[index].guardpage =(void *)(intptr_t)((uint64_t)(Class[index].guardpage)
-                + ((uint64_t)ret - old_ret));
-        Class[index].current =(void *)(intptr_t)((uint64_t)(Class[index].current)
-                + ((uint64_t)ret - old_ret));
+        ret = (void *)(intptr_t)(((intptr_t)ret + (intptr_t)(Class[index].size) - 1) &
+                ~((intptr_t)(Class[index].size) - 1));
+        Class[index].guardpage =(void *)(intptr_t)((intptr_t)(Class[index].guardpage)
+                + ((intptr_t)ret - old_ret));
+        Class[index].current =(void *)(intptr_t)((intptr_t)(Class[index].current)
+                + ((intptr_t)ret - old_ret));
     }
 
     return ret;
